Added new attribute `isMoved`to FileEntity and added new methods `getSecondary()`, `getPrimaryAndSecondary()`, `getDuplicates()`, `getAllFiles()` to MediaEntity
