+++ lib/steps/step_06_move_files/services/media_entity_moving_service.dart
@@ -426,30 +426,6 @@ class _Semaphore {
   }
 }
 
-// (resto del archivo sin cambios)
-
-
-/// Base class for MediaEntity moving strategies (unchanged public API)
-abstract class MediaEntityMovingStrategy {
-  const MediaEntityMovingStrategy();
-
-  String get name;
-  bool get createsShortcuts;
-  bool get createsDuplicates;
-
-  Stream<MediaEntityMovingResult> processMediaEntity(
-    final MediaEntity entity,
-    final MovingContext context,
-  );
-
-  Future<List<MediaEntityMovingResult>> finalize(
-    final MovingContext context,
-    final List<MediaEntity> processedEntities,
-  ) async => [];
-
-  void validateContext(final MovingContext context) {}
-}
-
 /// Represents a single file moving operation
 class MediaEntityMovingOperation {
   const MediaEntityMovingOperation({
@@ -523,6 +499,27 @@ class MediaEntityMovingResult {
   bool get isFailure => !success;
 }
 
+/// Base class for MediaEntity moving strategies (unchanged public API)
+abstract class MediaEntityMovingStrategy {
+  const MediaEntityMovingStrategy();
+
+  String get name;
+  bool get createsShortcuts;
+  bool get createsDuplicates;
+
+  Stream<MediaEntityMovingResult> processMediaEntity(
+    final MediaEntity entity,
+    final MovingContext context,
+  );
+
+  Future<List<MediaEntityMovingResult>> finalize(
+    final MovingContext context,
+    final List<MediaEntity> processedEntities,
+  ) async => [];
+
+  void validateContext(final MovingContext context) {}
+}
+
 /// Factory to create strategy by AlbumBehavior
 class MediaEntityMovingStrategyFactory {
   const MediaEntityMovingStrategyFactory(
@@ -538,19 +535,11 @@ class MediaEntityMovingStrategyFactory {
   MediaEntityMovingStrategy createStrategy(final AlbumBehavior albumBehavior) {
     switch (albumBehavior) {
       case AlbumBehavior.shortcut:
-        return ShortcutMovingStrategy(
-          _fileService,
-          _pathService,
-          _symlinkService,
-        );
+        return ShortcutMovingStrategy(_fileService, _pathService, _symlinkService);
       case AlbumBehavior.duplicateCopy:
         return DuplicateCopyMovingStrategy(_fileService, _pathService);
       case AlbumBehavior.reverseShortcut:
-        return ReverseShortcutMovingStrategy(
-          _fileService,
-          _pathService,
-          _symlinkService,
-        );
+        return ReverseShortcutMovingStrategy(_fileService, _pathService, _symlinkService);
       case AlbumBehavior.json:
         return JsonMovingStrategy(_fileService, _pathService);
       case AlbumBehavior.nothing:
@@ -559,140 +548,3 @@ class MediaEntityMovingStrategyFactory {
   }
 }
 
-/// ─────────────────────────────────────────────────────────────────────────
-/// Common helpers to avoid code duplication across strategies
-/// (centralized in this service module so strategies can import and reuse them)
-/// ─────────────────────────────────────────────────────────────────────────
-class MovingStrategyUtils {
-  const MovingStrategyUtils._();
-
-  /// Generate ALL_PHOTOS target directory (date-structured if needed).
-  static Directory allPhotosDir(
-    final PathGeneratorService pathService,
-    final MediaEntity entity,
-    final MovingContext context,
-  ) => pathService.generateTargetDirectory(
-      null,
-      entity.dateTaken,
-      context,
-      isPartnerShared: entity.partnerShared,
-    );
-
-  /// Generate Albums/<albumName> target directory (date-structured if needed).
-  static Directory albumDir(
-    final PathGeneratorService pathService,
-    final String albumName,
-    final MediaEntity entity,
-    final MovingContext context,
-  ) => pathService.generateTargetDirectory(
-      albumName,
-      entity.dateTaken,
-      context,
-      isPartnerShared: entity.partnerShared,
-    );
-
-  /// Returns true if 'child' path equals or is a subpath of 'parent'.
-  static bool isSubPath(final String child, final String parent) {
-    final String c = child.replaceAll('\\', '/');
-    final String p = parent.replaceAll('\\', '/');
-    return c == p || c.startsWith('$p/');
-  }
-
-  /// Returns the directory (without trailing slash) of a path, handling both separators.
-  static String dirOf(final String p) {
-    final normalized = p.replaceAll('\\', '/');
-    final idx = normalized.lastIndexOf('/');
-    return idx < 0 ? '' : normalized.substring(0, idx);
-  }
-
-  /// Infer album name for a given file source directory using albumsMap metadata.
-  /// Returns null if no album matches.
-  static String? inferAlbumForSourceDir(
-    final MediaEntity entity,
-    final String fileSourceDir,
-  ) {
-    for (final entry in entity.albumsMap.entries) {
-      for (final src in entry.value.sourceDirectories) {
-        if (isSubPath(fileSourceDir, src)) return entry.key;
-      }
-    }
-    return entity.albumNames.isNotEmpty ? entity.albumNames.first : null;
-  }
-
-  /// Compute the list of album names a given file (by its source directory) belonged to.
-  static List<String> albumsForFile(
-    final MediaEntity entity,
-    final FileEntity file,
-  ) {
-    final fileDir = dirOf(file.sourcePath);
-    final List<String> result = <String>[];
-    for (final entry in entity.albumsMap.entries) {
-      for (final src in entry.value.sourceDirectories) {
-        if (isSubPath(fileDir, src)) {
-          result.add(entry.key);
-          break;
-        }
-      }
-    }
-    return result;
-  }
-
-  /// Predicate: whether [file] belonged to the given [albumName] according to sourceDirectories.
-  static bool fileBelongsToAlbum(
-    final MediaEntity entity,
-    final FileEntity file,
-    final String albumName,
-  ) {
-    final info = entity.albumsMap[albumName];
-    if (info == null || info.sourceDirectories.isEmpty) return false;
-    final fileDir = dirOf(file.sourcePath);
-    for (final src in info.sourceDirectories) {
-      if (isSubPath(fileDir, src)) return true;
-    }
-    return false;
-  }
-
-  /// Create a symlink to [target] inside [dir] and try to rename it to [preferredBasename].
-  /// On name collision, appends " (n)" before extension.
-  static Future<File> createSymlinkWithPreferredName(
-    final SymlinkService symlinkService,
-    final Directory dir,
-    final File target,
-    final String preferredBasename,
-  ) async {
-    final File link = await symlinkService.createSymlink(dir, target);
-    final String currentBase = link.uri.pathSegments.last;
-    if (currentBase == preferredBasename) return link;
-
-    final String finalBasename = _resolveUniqueBasename(dir, preferredBasename);
-    final String desiredPath = '${dir.path}/$finalBasename';
-    try {
-      return await link.rename(desiredPath);
-    } catch (_) {
-      return link;
-    }
-  }
-
-  static String _resolveUniqueBasename(final Directory dir, final String base) {
-    final int dot = base.lastIndexOf('.');
-    final String stem = dot > 0 ? base.substring(0, dot) : base;
-    final String ext = dot > 0 ? base.substring(dot) : '';
-    String candidate = base;
-    int idx = 1;
-    while (_existsAny('${dir.path}/$candidate')) {
-      candidate = '$stem ($idx)$ext';
-      idx++;
-    }
-    return candidate;
-  }
-
-  static bool _existsAny(final String fullPath) {
-    try {
-      return File(fullPath).existsSync() ||
-          Link(fullPath).existsSync() ||
-          Directory(fullPath).existsSync();
-    } catch (_) {
-      return false;
-    }
-  }
-}
